protection against too large ratios between MC and Data events; changed variables from int to short

+++ ConfigParser/ParserBase.hh
@@ -72,8 +72,8 @@ public:
   const int noOfThreads() const {return _noOfThreads;}
   const int noOfClients() const {return _noOfClients;}
   const int serverPort() const {return _serverPort;}
-  const int ratioMcToData() const {return _ratioMcToData;}
-  const int ratioTruthToMc() const {return _ratioTruthToMc;}
+  const short ratioMcToData() const {return _ratioMcToData;}
+  const short ratioTruthToMc() const {return _ratioTruthToMc;}
   const int evoPopulation() const {return _evoPopulation;}
   const int evoIterations() const {return _evoIterations;}
   const double evoRatioOfModParams() const {return _evoRatioOfModParams;}
@@ -154,8 +154,8 @@ protected:
   int _noOfThreads;
   int _noOfClients;
   int _serverPort;
-  int _ratioMcToData;
-  int _ratioTruthToMc;
+  short _ratioMcToData;
+  short _ratioTruthToMc;
   int _evoIterations;
   int _evoPopulation;
   double _evoRatioOfModParams;
